+++ path_planner_c/old_path_planner.c
@@ -1,0 +1,267 @@
+
+/*
+* AstroTinker Bot (AB): Task 1B Path Planner
+*
+* This program computes the valid path from the start point to the end point.
+* Make sure you don't change anything outside the "Add your code here" section.
+* Updated memory addresses for Task 2B
+*/
+
+#include <stdlib.h>
+#include <stdbool.h>
+#include <stdint.h>
+
+#ifdef __linux__ // for host pc
+
+    #include <stdio.h>
+
+    void _put_byte(char c) { putchar(c); }
+
+    void _put_str(char *str) {
+        while (*str) {
+            _put_byte(*str++);
+        }
+    }
+
+    void print_output(uint8_t num) {
+        if (num == 0) {
+            putchar('0'); // if the number is 0, directly print '0'
+            _put_byte('\n');
+            return;
+        }
+
+        if (num < 0) {
+            putchar('-'); // print the negative sign for negative numbers
+            num = -num;   // make the number positive for easier processing
+        }
+
+        // convert the integer to a string
+        char buffer[20]; // assuming a 32-bit integer, the maximum number of digits is 10 (plus sign and null terminator)
+        uint8_t index = 0;
+
+        while (num > 0) {
+            buffer[index++] = '0' + num % 10; // convert the last digit to its character representation
+            num /= 10;                        // move to the next digit
+        }
+
+        // print the characters in reverse order (from right to left)
+        while (index > 0) { putchar(buffer[--index]); }
+        _put_byte('\n');
+    }
+
+    void _put_value(uint8_t val) { print_output(val); }
+
+#else  // for the test device
+
+    void _put_value(uint8_t val) { }
+    void _put_str(char *str) { }
+
+#endif
+
+/*
+Functions Usage
+
+instead of using printf() function for debugging,
+use the below function calls to print a number, string or a newline
+
+for newline: _put_byte('\n');
+for string:  _put_str("your string here");
+for number:  _put_value(your_number_here);
+
+Examples:
+        _put_value(START_POINT);
+        _put_value(END_POINT);
+        _put_str("Hello World!");
+        _put_byte('\n');
+
+*/
+
+//###################functions###################
+// Number of vertices in the graph
+#define V 30
+#define INT_MAX 0x7fffffff
+
+
+// helper function to find the vertex with minimum distance value and not visited
+int minDistance(int dist[], bool sptSet[])
+{
+	int min = INT_MAX, min_index;
+
+	for (int v = 0; v < V; v++)
+		if (sptSet[v] == false && dist[v] <= min)
+			min = dist[v], min_index = v;
+
+	return min_index;
+}
+
+//recursive function to find the path from the parent array
+void findPath(int currentVertex, int parents[], uint8_t path_planned[], uint8_t *idx)
+{
+
+    if (currentVertex == -1) {
+        return;
+    }
+    findPath(parents[currentVertex], parents, path_planned, idx);
+    
+    path_planned[(*idx)++] = currentVertex;
+}
+
+
+// function to find the shortest path from source to destination using Dijkstra's algorithm
+// refered from geekforgeeks
+void dijkstra(uint8_t graph[V][V], int src, int dest, uint8_t path_planned[], uint8_t *idx)
+{
+	int dist[V]; 
+    int parent[V];
+
+	bool sptSet[V]; 
+	for (int i = 0; i < V; i++)
+		dist[i] = INT_MAX, sptSet[i] = false;
+
+	dist[src] = 0;
+    parent[src] = -1;
+
+	for (int count = 0; count < V - 1; count++) {
+		int u = minDistance(dist, sptSet);
+
+        //break if the destination is reached
+        if (u == dest)
+            break;
+        //mark the vertex as visited
+		sptSet[u] = true;
+
+        // updating the distance of the adjacent unvisited vertices
+		for (int v = 0; v < V; v++)
+			if (!sptSet[v] && graph[u][v]   // if the vertex is not visited (value of current v has never been the value of u) 
+                                            //and there is an edge between u and v
+				&& dist[u] != INT_MAX  // if the distance of u is not infinity
+				&& dist[u] + graph[u][v] < dist[v]) // sum of distance to u and edge weight of v-u is 
+                                                    //less than the distance to v (v is not visited)
+                {
+                    dist[v] = dist[u] + graph[u][v]; //update the distance of v     
+                    parent[v] = u; //update the parent of v
+                }
+	}
+
+    findPath(dest, parent, path_planned, idx);
+}
+//#############################################
+
+
+
+// main function
+int main(int argc, char const *argv[]) {
+
+    #ifdef __linux__
+
+        const uint8_t START_POINT   = atoi(argv[1]);
+        const uint8_t END_POINT     = atoi(argv[2]);
+        uint8_t NODE_POINT          = 0;
+        uint8_t CPU_DONE            = 0;
+
+    #else
+        // Address value of variables are updated for RISC-V Implementation
+        #define START_POINT         (* (volatile uint8_t * ) 0x02000000)
+        #define END_POINT           (* (volatile uint8_t * ) 0x02000004)
+        #define NODE_POINT          (* (volatile uint8_t * ) 0x02000008)
+        #define CPU_DONE            (* (volatile uint8_t * ) 0x0200000c)
+
+    #endif
+
+    // array to store the planned path
+    uint8_t path_planned[32];
+    // index to keep track of the path_planned array
+    uint8_t idx = 0;
+
+    // ############# Add your code here #############
+    //################### my code###################
+    //adjacency matrix of the graph
+    // adjacency matrix of the graph
+    uint8_t map[30][30] = {
+        // row 0
+        {0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 1
+        {1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
+        // row 2
+        {0,1,0,1,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 3
+        {0,0,1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0},
+        // row 4
+        {0,0,0,1,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 5
+        {0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 6
+        {0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 7
+        {0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 8
+        {0,0,1,0,0,0,0,1,0,1,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 9
+        {0,0,0,0,0,0,0,0,1,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 10
+        {0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 11
+        {0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 12
+        {0,0,0,0,0,0,0,0,1,0,0,0,0,1,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0},
+        // row 13
+        {0,0,0,0,0,0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 14
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 15
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 16
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0},
+        // row 17
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0},
+        // row 18
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,0},
+        // row 19
+        {0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,0,0,0},
+        // row 20
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,1,0,0,1,0,0,0,0,1},
+        // row 21
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,1,1,0,0,0,0,0,0},
+        // row 22
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0},
+        // row 23
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0},
+        // row 24
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,1,0,0,0,0},
+        // row 25
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,1,0,0,0},
+        // row 26
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,1,1,0},
+        // row 27
+        {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0},
+        // row 28
+        {0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,1},
+        // row 29
+        {0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,1,0}
+    };
+    
+    dijkstra(map, START_POINT, END_POINT, path_planned,&idx);
+
+
+    // ##############################################
+
+    // the node values are written into data memory sequentially.
+    for (int i = 0; i < idx; ++i) {
+        NODE_POINT = path_planned[i];
+    }
+    // Path Planning Computation Done Flag
+    CPU_DONE = 1;
+
+    #ifdef __linux__    // for host pc
+
+        _put_str("######### Planned Path #########\n");
+        for (int i = 0; i < idx; ++i) {
+            _put_value(path_planned[i]);
+        }
+        _put_str("################################\n");
+
+    #endif
+
+    return 0;
+}
+
