+++ simulation/modelsim/dump_txt.txt
@@ -1,0 +1,5 @@
+0
+1
+2
+8
+7
