+++ t2b_riscv_cpu_nativelink_simulation.rpt
@@ -18,6 +18,5 @@ Sim dir               :  simulation\modelsim
 
 Info: Starting NativeLink simulation with ModelSim-Altera software
 Sourced NativeLink script d:/intelfpga_lite/20.1/quartus/common/tcl/internal/nativelink/modelsim.tcl
-Warning: File t2b_riscv_cpu_run_msim_rtl_verilog.do already exists - backing up current file as t2b_riscv_cpu_run_msim_rtl_verilog.do.bak
 Info: Spawning ModelSim-Altera Simulation software 
 Info: NativeLink simulation flow was successful
